entity movement fixed

+++ JourneyCoreDisplay/Graphics/Drawing/DrawQueueItem.cs
@@ -6,9 +6,9 @@ namespace JourneyCoreLib.Drawing
     public class DrawQueueItem
     {
         public DrawPriority PriorityLevel { get; }
-        public Action<int, RenderWindow> Draw { get; }
+        public Action<float, RenderWindow> Draw { get; }
 
-        public DrawQueueItem(DrawPriority priorityLevel, Action<int, RenderWindow> drawingFunction)
+        public DrawQueueItem(DrawPriority priorityLevel, Action<float, RenderWindow> drawingFunction)
         {
             PriorityLevel = priorityLevel;
             Draw = drawingFunction;
